refactor: 받침 체크 파일 위치 변경, suffix 수정
